sample selector sidebar
